Refactors

+++ src/main/java/com/phones/phones/projection/CityTop.java
@@ -3,6 +3,6 @@ package com.phones.phones.projection;
 public interface CityTop {
 
     String getName();
-    Integer getQuantity();
+    Long getQuantity();
 
 }
